logs-service/requirements.txt
redis
websockets
asyncio
Flask==1.1.4 # it seems like Flask-sockets doesn't play well with 2.0+
Flask-Sockets==0.2.1
gevent
requests
python-decouple
